Remove old scripts not in use
We can use the pull / pr functionality in Lokalise
